Enabled debug postfix for library filenames
This will append a _d to all library filenames in Debug builds.
This is useful because on the MSVC toolchains Debug and Release
use different incompatible runtimes, so two sets of binaries need
to be kept around for a fully usable installation. Since SFML
currently saves all the binaries to the same directory during
INSTALL, the postfix resolves the name clash when installing
both Debug and Release configurations.