ImGui-SFML v2.2
This release is an accumulation small things which are important to release:
* ImGui >= 1.80 is now required. ImGui added imgui_tables.cpp and some
of the functions were deprecated. It's important to stay up-to-date
* Fixed Image/ImageButton and sf::RenderTexture problems (texture was
shown upside down)
* Removed an overload for Image/ImageButton which took sf::Texture and
sf::IntRect textureRect. You should use sf::Sprite now.
* Added sf::Cursor::Hand (thanks, @DarkContact)
* Restored C++03 support
* Fix MSVC warnings (thanks, @andrew-gresyk, @plutphil)

Also I'd like to thank @pinam45 for making CI for ImGui-SFML. Also
thanks to @sabidib, @dubgron and @ComicSansMS for improving/fixing
builds and CMake scripts.