Fix DLL builds
The problem boiled down to the fact that the examples added
imgui_demo.cpp as a source file without correctly defining IMGUI_API
while building it.

https://github.com/ocornut/imgui/blob/6eb2681c0988d8ccd82be5f0a6581a0e577bea86/imgui_demo.cpp#L13-L27

The ImGui maintainer recommends we always build imgui_demo.cpp so I
changed IMGUI_SFML_IMGUI_DEMO to ON by default so that most users
will get this. This option will potentially be removed in a future
release since it adds some additional complexity to the build for
margin gains.

The only reason you'd omit this is for binary size reasons when
building shared libs. The ImGui maintainer already recommends against
using shared library and on top of that I only measured a 300 kB binary
size increase from 1.4 MB to 1.7 MB.