Add unit tests
The tests fail to run when building SFML as a shared library because
GCC can't find libsfml-window.so.2.6 and they fail on Windows because
shared libraries just always suck on Windows. Disabling unit tests
for all shared libraries builds is a lot easier plus should not have
a negative impact on the utility of the tests.

The tests are disabled on macOS because there's some weird issue with
the linker not being about to find Freetype.framework. Not sure what
that's about so I'm just disabling the tests on macOS. Hopefully it
won't be a problem.